@@ -1,33 +1,17 @@
 part of flutter_bluetooth_serial;
 
-/// Represents information about the device. Could be out-of-date. // @TODO . add updating the info via some fn
 class BluetoothDevice {
-  /// Broadcasted friendly name of the device.
   final String? name;
-
-  /// MAC address of the device or identificator for platform system (if MAC addresses are prohibited).
   final String address;
-
-  /// Type of the device (Bluetooth standard type).
   final BluetoothDeviceType type;
-
-  /// Class of the device.
-  //final BluetoothClass bluetoothClass // @TODO . !BluetoothClass!
-
-  /// Describes is device connected.
   final bool isConnected;
-
-  /// Bonding state of the device.
   final BluetoothBondState bondState;
 
-  /// Tells whether the device is bonded (ready to secure connect).
   @Deprecated('Use `isBonded` instead')
   bool get bonded => bondState.isBonded;
 
-  /// Tells whether the device is bonded (ready to secure connect).
   bool get isBonded => bondState.isBonded;
 
-  /// Construct `BluetoothDevice` with given values.
   const BluetoothDevice({
     this.name,
     required this.address,
@@ -36,9 +20,6 @@ class BluetoothDevice {
     this.bondState = BluetoothBondState.unknown,
   });
 
-  /// Creates `BluetoothDevice` from map.
-  ///
-  /// Internally used to receive the object from platform code.
   factory BluetoothDevice.fromMap(Map map) {
     return BluetoothDevice(
       name: map["name"],
@@ -53,23 +34,40 @@ class BluetoothDevice {
     );
   }
 
-  /// Creates map from `BluetoothDevice`.
   Map<String, dynamic> toMap() => {
-        "name": this.name,
-        "address": this.address,
-        "type": this.type.toUnderlyingValue(),
-        "isConnected": this.isConnected,
-        "bondState": this.bondState.toUnderlyingValue(),
+        "name": name,
+        "address": address,
+        "type": type.toUnderlyingValue(),
+        "isConnected": isConnected,
+        "bondState": bondState.toUnderlyingValue(),
       };
 
-  /// Compares for equality of this and other `BluetoothDevice`.
-  ///
-  /// In fact, only `address` is compared, since this is most important
-  /// and unchangable information that identifies each device.
-  operator ==(Object other) {
-    return other is BluetoothDevice && other.address == this.address;
+  BluetoothDevice copyWith({
+    String? name,
+    String? address,
+    BluetoothDeviceType? type,
+    bool? isConnected,
+    BluetoothBondState? bondState,
+  }) {
+    return BluetoothDevice(
+      name: name ?? this.name,
+      address: address ?? this.address,
+      type: type ?? this.type,
+      isConnected: isConnected ?? this.isConnected,
+      bondState: bondState ?? this.bondState,
+    );
+  }
+
+  @override
+  bool operator ==(Object other) {
+    return other is BluetoothDevice && other.address == address;
   }
 
   @override
   int get hashCode => address.hashCode;
+
+  @override
+  String toString() {
+    return 'BluetoothDevice{name: $name, address: $address, type: $type, isConnected: $isConnected, bondState: $bondState}';
+  }
 }
